Created DataBaseAPI, WebSocketController classes with public static methods
