--- src/controllers/webSocketController.ts
@@ -1,9 +1,0 @@
-import ClientConnectionService from '../services/ClientConnectionService';
-import { WebSocket, WebSocketServer } from 'ws';
-
-export function clientConnectionController(
-  clientSocket: WebSocket,
-  webSocketServer: WebSocketServer
-) {
-  ClientConnectionService.clientConnection(clientSocket, webSocketServer);
-}
